refactor: width 부분 hook으로 분리
